requirements.txt
requests>=2.31
python-dotenv>=1.0
cryptography>=42.0
pandas>=2.2
